# backend/admin-gin/api/openapi/swagger.yaml
basePath: /
definitions:
  auth.LoginReq:
    properties:
      password:
        description: 密码
        example: "123456"
        type: string
      username:
        description: 用户名
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  auth.LoginRes:
    properties:
      access_token:
        description: 访问令牌
        type: string
      expires_at:
        description: 过期时间
        type: integer
      refresh_token:
        description: 刷新令牌
        type: string
      uid:
        description: 用户UID
        type: string
    type: object
  auth.RefreshReq:
    properties:
      refresh_token:
        description: 刷新令牌
        type: string
    required:
    - refresh_token
    type: object
  auth.RegisterReq:
    properties:
      password:
        description: 密码
        example: "123456"
        type: string
      username:
        description: 用户名
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  http.HttpResponse-any:
    properties:
      code:
        description: 'code: HTTP 状态码（与 net/http 保持一致）'
        example: 200
        type: integer
      data:
        description: 'data: 响应数据（可以为空）'
      message:
        description: 'message: 响应描述（默认可用 internal/pkg/http.StatusText(code)）'
        example: 成功
        type: string
    type: object
  http.HttpResponse-auth_LoginRes:
    properties:
      code:
        description: 'code: HTTP 状态码（与 net/http 保持一致）'
        example: 200
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/auth.LoginRes'
        description: 'data: 响应数据（可以为空）'
      message:
        description: 'message: 响应描述（默认可用 internal/pkg/http.StatusText(code)）'
        example: 成功
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Mall API 服务接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Mall API
  version: "1.0"
paths:
  /admin/auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名密码登录，成功后返回 token 对
      parameters:
      - description: 登录信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: 失败
          schema:
            $ref: '#/definitions/http.HttpResponse-any'
      summary: 用户登录
      tags:
      - Auth
  /admin/auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用 refresh token 刷新访问 token，成功后返回新的 token 对
      parameters:
      - description: 刷新 token 请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshReq'
      produces:
      - application/json
      responses:
        "200":
          description: 失败
          schema:
            $ref: '#/definitions/http.HttpResponse-any'
      summary: 刷新 Token
      tags:
      - Auth
  /admin/auth/register:
    post:
      consumes:
      - application/json
      description: 使用用户名密码进行注册
      parameters:
      - description: 注册信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: 失败
          schema:
            $ref: '#/definitions/http.HttpResponse-any'
      summary: 用户注册
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"